Updates navigation link target in scripts.js
Instead of document.location.href, the location that's routed is
window.location.href. Hopefully that will fix the navigation problem of
the launched website.

The navigation problem being: when a navbar button is clicked, the page
is routed to murgray.com/<file>.html_files instead of
murgray.com/html_files/<file>.html.